ml/ohp_form_pipeline/requirements.txt
# Core
numpy>=1.24
scipy>=1.10
pandas>=2.0
pyyaml>=6.0
tqdm>=4.65

# Computer Vision
opencv-python>=4.8
mediapipe>=0.10

# Deep Learning (optional but recommended)
torch>=2.1
torchvision>=0.16
transformers>=4.37
peft>=0.7
einops>=0.7

# Signal processing / ML
scikit-learn>=1.3
hdbscan>=0.8.33
umap-learn>=0.5.5
tslearn>=0.6

# Visualization
matplotlib>=3.7
plotly>=5.15

# Utilities
Pillow>=10.0

# PDF extraction (for training data from S&C manual)
PyMuPDF>=1.23

# Experiment tracking & tracing
wandb>=0.16
weave>=0.50
python-dotenv>=1.0
